+++ Sims1WidescreenPatcher.Core/Models/ScalePanelBackJob.cs
@@ -12,7 +12,7 @@ public class ScalePanelBackJob : BaseImageProcessingJob
         var left = image.Clone(0, 0, 286, 100);
         var middle = image.Clone(left.Width, 0, 500, 100);
         var right = image.Clone(left.Width + middle.Width, 0, 18, 100);
-        middle.Resize(new MagickGeometry(Width - left.Width - right.Width, Height) { IgnoreAspectRatio = true });
+        middle.Resize(new MagickGeometry((int)Width - left.Width - right.Width, (int)Height) { IgnoreAspectRatio = true });
         left.Page = new MagickGeometry("+0+0");
         middle.Page = new MagickGeometry($"+{left.Width}+0");
         right.Page = new MagickGeometry($"+{left.Width + middle.Width}+0");
